Delete tests/generate and move the content to tests/test_cache.py
